support 0 size file

+++ src/renderer/logic/downloader.js
@@ -86,6 +86,20 @@ async function downloadFileTemp(fileObj, silent = false) {
   })
 }
 
+async function downloadZeroSizeFile(fileObj, silent = false) {
+  const originalFileName = path.basename(fileObj.fullpath)
+
+  // Delete temp file
+  if (fs.existsSync(fileObj.fullpath)) {
+    try {
+      fs.unlinkSync(fileObj.fullpath)
+    } catch (e) {
+      Logger.error('Delete local file: Cannot delete', e.message)
+    }
+  }
+  fs.closeSync(fs.openSync(fileObj.fullpath, 'w'))
+}
+
 // Will download ALL the files from remote
 // If file already exists on local, decide if needs to be checked.
 async function _downloadAllFiles() {
@@ -175,17 +189,20 @@ async function _downloadAllFiles() {
           Logger.error(err)
         })
       try {
-        const tempPath = await downloadFileTemp(item)
-        if (localExists) {
-          try {
-            fs.unlinkSync(item.fullpath)
-          } catch (e) {}
+        if (item.size !== 0) {
+          const tempPath = await downloadFileTemp(item)
+          if (localExists) {
+            try {
+              fs.unlinkSync(item.fullpath)
+            } catch (e) {}
+          }
+          // fs.renameSync gives a "EXDEV: cross-device link not permitted"
+          // when application and local folder are not in the same partition
+          fs.copyFileSync(tempPath, item.fullpath)
+          fs.unlinkSync(tempPath)
+        } else {
+          await downloadZeroSizeFile(item)
         }
-        // fs.renameSync gives a "EXDEV: cross-device link not permitted"
-        // when application and local folder are not in the same partition
-        fs.copyFileSync(tempPath, item.fullpath)
-        fs.unlinkSync(tempPath)
-
         await Sync.setModifiedTime(item.fullpath, item.created_at)
 
         analytics
@@ -225,16 +242,26 @@ async function _downloadAllFiles() {
         ].find(obj => obj === err.message)
         if (isError) {
           Logger.error(err.message)
-          await Database.dbFiles.remove({ key: item.fullpath })
-          continue
         }
-        // continue
+        await Database.dbFiles.remove({ key: item.fullpath })
+        continue
       }
     } else if (uploadAndReplace) {
+      const stat = Tree.getStat(item.fullpath)
       const storj = await getEnvironment()
-      await Uploader.uploadFile(storj, item.fullpath, currentFiles, totalFiles)
+      await Uploader.uploadFile(
+        storj,
+        item.fullpath,
+        currentFiles,
+        totalFiles,
+        item
+      )
+
       // continue
     } else {
+      if (item.size === 0) {
+        continue
+      }
       // Check if should download to ensure file
       const shouldEnsureFile = Math.floor(Math.random() * 33 + 1) % 33 === 0
